Add files via upload
This project Investigates correlation and predictive relationships within Nasdaq 100 Tech Stocks and QQQ ETF to determine individual stocks and stock groups that drive the movement of the index. In the ever-evolving landscape of financial markets, understanding the intricate relationships between individual stocks and exchange-traded funds (ETFs) is crucial for informed decision-making. This machine learning (ML) project investigates the relationships between the stock prices of the top 10 and bottom 10 companies listed on the Nasdaq 100 index and the corresponding prices of the QQQ ETF to find out which stocks drive the movement of the QQQ ETF.